Implemented access keys and privileges.
Clients now require an access key specified in the conf file. This
access key can be either a mailbox key which allows access to get and
remove messages from a single mailbox or an API key which allows full
access to the system.

Mailbox access keys are generated when the mailbox is registered.

API access keys can be created from the console with:

conduit server access