Script deployment grouping
Scripts are now grouped into a "deployment" object which stores the
script body and deployment details. Messages now simply link the
deployments to mailboxes.

This will reduce the storage overhead by not storing a script's body
once per mailbox. It will now be stored only once for the entire deploy
and linked to each mailbox.

Scripts no longer are removed from the database, instead they are marked
as "deleted". Messages marked as deleted are ignored when pulling from
the queue.

Deployment responses were added. Clients can now respond with
information using the "$" function within a script. These responses are
tied to the deployment.

CLI Send command has been reworked to pull back deployment responses as
well as continously poll the server until the deployment is completed.
A timeout may need to be added to this functionality at a later time.