Version 0.3-beta.3
Deregistering a non existing mailbox caused an err

Request errors were not being checked on the client's deregister
function. This was changed to return the error appropriately.

Fixed asset attachment

The multipart writer was not closing before the body was being written
out this was causing it to miss the last boundary. Added a call to
Close() before writing.

Added tests for asset attachment and downloading in the cmd package.

Added version 2 db fields to CreateDB

Changed identifiers to be a hex string instead of hyphen seperated

Added $file.eachLine engine function

Added error checking for the conduit agent command

If 'conduit agent' was ran without agent_host in the config file it
would simply die. Added error checking to see if that field exists
before attempting to start. If it does not it responds with the agent
command help.