Version 0.4-beta.2
Server stability changes

The poll time for the server is now 100 - 600 seconds chosen at random.
This is an increase over 100 seconds in order to space out requests. The
randomized factor was added to help distribute client reconnects over
time.

Deployments are now staggered. When a script is deployed it will be
deployed to a 20 clients per second. This delay was added to help
stagger the requests slightly to help the server from getting bogged
down with too many requests at once.

Client stability changes

Client reconnects are now delayed by 1-3 seconds. This was added to
prevent servers from rapidly reconnecting. The randomness was added to
help distribute reconnects from large amounts of clients at the same
time.

Error back off was changed from n^2 to n^2 + (0 to n). When the client
receives an error from the server the time it waits to reconnect is
based on the number of attempts it has made (to a maximum of 15).
Another factor of 0 to n was added to help space out reconnects in
situations such as the server being restarted.

Logging reduction

Several log messages were moved into debug messages so they will not
appear in the log unless the -d flag is being used.

Added $system.exit() function to the scripting engine which will cause
conduit to shutdown. This was added to make it easier to upgrade conduit
using a script.